Fixed issues with imu-monocular mode crashing
There was a bunch of issues I was running into with crashing for
different reasons every time, so I reverted to an old branch of
orb_slam3 and made some small changes only. I removed a line in
optimizer.cc that had an assert for mit->second>=3 which was always
failing. Apparently, in [this](https://github.com/UZ-SLAMLab/ORB_SLAM3/issues/145)
github issue, compiling in Release mode instead of Debug mode fixed this
issue. It seems that was the case because the compiler would ignore
these asserts in Release mode but not in Debug mode. So I could continue
to use gdb, I removed this assert and the code works very smoothly now,
consistently building maps all the way to the end of the ros bags that I
have.